Refactore baseline fs2 client bench
