+++ benchmarks/fs2/src/main/scala/laserdisc/fs2/parallel/ParallelLoadLaserdiscCatsBaselineBench.scala
@@ -8,7 +8,7 @@ import cats.effect.{Blocker, ContextShift, IO, Timer}
 import cats.syntax.flatMap._
 import laserdisc.auto._
 import laserdisc.fs2.parallel.SetUpLaserdiscCatsBaseline.LaserdiscCatsBaselineSetUp
-import laserdisc.fs2.parallel.baseline.RedisBytesChannel
+import laserdisc.fs2.parallel.baseline.BenchRedisBaselineChannel
 import laserdisc.fs2.parallel.testcases.TestCasesLaserdiscBaseline
 import log.effect.fs2.SyncLogWriter
 import log.effect.{LogLevels, LogWriter}
@@ -52,7 +52,7 @@ object SetUpLaserdiscCatsBaseline {
 
     val resource = Blocker[IO] evalMap { bl =>
       RedisAddress("localhost", 6379).toInetSocketAddress[IO] map { address =>
-        RedisBytesChannel[IO](address, writeTimeout = Some(10.seconds), readMaxBytes = 8 * 1024 * 1024)(bl)
+        BenchRedisBaselineChannel[IO](address, writeTimeout = Some(10.seconds), readMaxBytes = 8 * 1024 * 1024)(bl)
       }
     }
